third_party/WebKit/LayoutTests/fast/events/mutation-during-insert-before-expected.txt

Ensures that insertBefore() throws an exception if mutation even handler does something wrong

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS newparent.insertBefore(child, beforeChild); threw exception HierarchyRequestError: Failed to execute 'insertBefore' on 'ContainerNode': The new child element contains the parent..
PASS successfullyParsed is true

TEST COMPLETE
